Move temp into cache_retry
